+++ src/filters/BlendImage.ts
@@ -1,7 +1,7 @@
 import { Image } from '../shapes/Image';
 import type { TClassProperties } from '../typedefs';
 import { createCanvasElement } from '../util/misc/dom';
-import { AbstractBaseFilter } from './BaseFilter';
+import { BaseFilter } from './BaseFilter';
 import type {
   T2DPipelineState,
   TWebGLPipelineState,
@@ -9,6 +9,9 @@ import type {
 } from './typedefs';
 import { WebGLFilterBackend } from './WebGLFilterBackend';
 import { classRegistry } from '../ClassRegistry';
+import { fragmentSource } from './blendImage.shaders';
+
+export type TBlendImageMode = 'multiply' | 'mask';
 
 export const blendImageDefaultValues: Partial<TClassProperties<BlendImage>> = {
   mode: 'multiply',
@@ -24,36 +27,6 @@ export const blendImageDefaultValues: Partial<TClassProperties<BlendImage>> = {
       gl_Position = vec4(aPosition * 2.0 - 1.0, 0.0, 1.0);
     }
     `,
-  fragmentSource: {
-    multiply: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      uniform sampler2D uImage;
-      uniform vec4 uColor;
-      varying vec2 vTexCoord;
-      varying vec2 vTexCoord2;
-      void main() {
-        vec4 color = texture2D(uTexture, vTexCoord);
-        vec4 color2 = texture2D(uImage, vTexCoord2);
-        color.rgba *= color2.rgba;
-        gl_FragColor = color;
-      }
-      `,
-    mask: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      uniform sampler2D uImage;
-      uniform vec4 uColor;
-      varying vec2 vTexCoord;
-      varying vec2 vTexCoord2;
-      void main() {
-        vec4 color = texture2D(uTexture, vTexCoord);
-        vec4 color2 = texture2D(uImage, vTexCoord2);
-        color.a = color2.a;
-        gl_FragColor = color;
-      }
-      `,
-  },
 };
 
 /**
@@ -73,14 +46,14 @@ export const blendImageDefaultValues: Partial<TClassProperties<BlendImage>> = {
  * object.applyFilters();
  * canvas.renderAll();
  */
-export class BlendImage extends AbstractBaseFilter<Record<string, string>> {
+export class BlendImage extends BaseFilter {
   /**
    * Color to make the blend operation with. default to a reddish color since black or white
    * gives always strong result.
    **/
   declare image: Image;
 
-  declare mode: 'multiply' | 'mask';
+  declare mode: TBlendImageMode;
 
   /**
    * alpha value. represent the strength of the blend image operation.
@@ -95,7 +68,7 @@ export class BlendImage extends AbstractBaseFilter<Record<string, string>> {
   }
 
   getFragmentSource(): string {
-    return this.fragmentSource[this.mode];
+    return fragmentSource[this.mode];
   }
 
   applyToWebGL(options: TWebGLPipelineState) {
