+++ src/filters/Boilerplate.ts
@@ -3,17 +3,6 @@ import { BaseFilter } from './BaseFilter';
 import { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 
 export const myFilterDefaultValues: Partial<TClassProperties<MyFilter>> = {
-  fragmentSource: `
-    precision highp float;
-    uniform sampler2D uTexture;
-    uniform float uMyParameter;
-    varying vec2 vTexCoord;
-    void main() {
-      vec4 color = texture2D(uTexture, vTexCoord);
-      // add your gl code here
-      gl_FragColor = color;
-    }
-  `,
   myParameter: 0,
   mainParameter: 'myParameter',
 };
@@ -39,6 +28,19 @@ export class MyFilter extends BaseFilter {
 
   static defaults = myFilterDefaultValues;
 
+  getFragmentSource() {
+    return `
+      precision highp float;
+        uniform sampler2D uTexture;
+        uniform float uMyParameter;
+        varying vec2 vTexCoord;
+        void main() {
+          vec4 color = texture2D(uTexture, vTexCoord);
+          // add your gl code here
+          gl_FragColor = color;
+        }
+      `;
+  }
   /**
    * Apply the MyFilter operation to a Uint8ClampedArray representing the pixels of an image.
    *
