+++ src/filters/Brightness.ts
@@ -2,19 +2,8 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
-
+import { fragmentSource } from './brightness.shaders';
 export const brightnessDefaultValues: Partial<TClassProperties<Brightness>> = {
-  fragmentSource: `
-    precision highp float;
-    uniform sampler2D uTexture;
-    uniform float uBrightness;
-    varying vec2 vTexCoord;
-    void main() {
-      vec4 color = texture2D(uTexture, vTexCoord);
-      color.rgb += uBrightness;
-      gl_FragColor = color;
-    }
-  `,
   brightness: 0,
   mainParameter: 'brightness',
 };
@@ -39,6 +28,10 @@ export class Brightness extends BaseFilter {
   declare brightness: number;
 
   static defaults = brightnessDefaultValues;
+
+  getFragmentSource() {
+    return fragmentSource;
+  }
 
   /**
    * Apply the Brightness operation to a Uint8ClampedArray representing the pixels of an image.
