+++ src/filters/ColorMatrix.ts
@@ -2,21 +2,9 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
-
+import { fragmentSource } from './colorMatrix.shader';
 export const colorMatrixDefaultValues: Partial<TClassProperties<ColorMatrix>> =
   {
-    fragmentSource: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      varying vec2 vTexCoord;
-      uniform mat4 uColorMatrix;
-      uniform vec4 uConstants;
-      void main() {
-        vec4 color = texture2D(uTexture, vTexCoord);
-        color *= uColorMatrix;
-        color += uConstants;
-        gl_FragColor = color;
-      }`,
     matrix: [1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0],
     mainParameter: 'matrix',
     colorsOnly: true,
@@ -65,6 +53,10 @@ export class ColorMatrix extends BaseFilter {
       this.matrix = [...matrix];
     }
     Object.assign(this, options);
+  }
+
+  getFragmentSource(): string {
+    return fragmentSource;
   }
 
   /**
