+++ src/filters/Gamma.ts
@@ -1,26 +1,12 @@
 import type { TClassProperties } from '../typedefs';
-import { BaseFilter, BaseFilterOptions } from './BaseFilter';
+import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
-
+import { fragmentSource } from './gamma.shaders';
+import {BaseFilterOptions} from '../../dist/src/filters/BaseFilter';
 export type GammaInput = [number, number, number];
 
 export const gammaDefaultValues: Partial<TClassProperties<Gamma>> = {
-  fragmentSource: `
-    precision highp float;
-    uniform sampler2D uTexture;
-    uniform vec3 uGamma;
-    varying vec2 vTexCoord;
-    void main() {
-      vec4 color = texture2D(uTexture, vTexCoord);
-      vec3 correction = (1.0 / uGamma);
-      color.r = pow(color.r, correction.r);
-      color.g = pow(color.g, correction.g);
-      color.b = pow(color.b, correction.b);
-      gl_FragColor = color;
-      gl_FragColor.rgb *= color.a;
-    }
-  `,
   mainParameter: 'gamma',
   gamma: [1, 1, 1],
 };
@@ -48,6 +34,10 @@ export class Gamma extends BaseFilter {
   };
 
   static defaults = gammaDefaultValues;
+
+  getFragmentSource() {
+    return fragmentSource;
+  }
 
   constructor({
     gamma,
