+++ src/filters/Grayscale.ts
@@ -1,43 +1,12 @@
 import type { TClassProperties } from '../typedefs';
-import { AbstractBaseFilter } from './BaseFilter';
+import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
+import { fragmentSource } from './grayscale.shaders';
+
+export type TGrayscaleMode = 'average' | 'lightness' | 'luminosity';
 
 export const grayscaleDefaultValues: Partial<TClassProperties<Grayscale>> = {
-  fragmentSource: {
-    average: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      varying vec2 vTexCoord;
-      void main() {
-        vec4 color = texture2D(uTexture, vTexCoord);
-        float average = (color.r + color.b + color.g) / 3.0;
-        gl_FragColor = vec4(average, average, average, color.a);
-      }
-      `,
-    lightness: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      uniform int uMode;
-      varying vec2 vTexCoord;
-      void main() {
-        vec4 col = texture2D(uTexture, vTexCoord);
-        float average = (max(max(col.r, col.g),col.b) + min(min(col.r, col.g),col.b)) / 2.0;
-        gl_FragColor = vec4(average, average, average, col.a);
-      }
-      `,
-    luminosity: `
-      precision highp float;
-      uniform sampler2D uTexture;
-      uniform int uMode;
-      varying vec2 vTexCoord;
-      void main() {
-        vec4 col = texture2D(uTexture, vTexCoord);
-        float average = 0.21 * col.r + 0.72 * col.g + 0.07 * col.b;
-        gl_FragColor = vec4(average, average, average, col.a);
-      }
-      `,
-  },
   mode: 'average',
   mainParameter: 'mode',
 };
@@ -49,8 +18,8 @@ export const grayscaleDefaultValues: Partial<TClassProperties<Grayscale>> = {
  * object.filters.push(filter);
  * object.applyFilters();
  */
-export class Grayscale extends AbstractBaseFilter<Record<string, string>> {
-  declare mode: 'average' | 'lightness' | 'luminosity';
+export class Grayscale extends BaseFilter {
+  declare mode: TGrayscaleMode;
 
   static defaults = grayscaleDefaultValues;
 
@@ -83,12 +52,8 @@ export class Grayscale extends AbstractBaseFilter<Record<string, string>> {
     }
   }
 
-  getCacheKey() {
-    return `${this.type}_${this.mode}`;
-  }
-
   getFragmentSource() {
-    return this.fragmentSource[this.mode];
+    return fragmentSource[this.mode];
   }
 
   /**
