+++ src/filters/blendImage.shaders.ts
@@ -1,0 +1,32 @@
+import { TBlendImageMode } from "./BlendImage";
+
+export const fragmentSource: Record<TBlendImageMode, string> = {
+  multiply: `
+    precision highp float;
+    uniform sampler2D uTexture;
+    uniform sampler2D uImage;
+    uniform vec4 uColor;
+    varying vec2 vTexCoord;
+    varying vec2 vTexCoord2;
+    void main() {
+      vec4 color = texture2D(uTexture, vTexCoord);
+      vec4 color2 = texture2D(uImage, vTexCoord2);
+      color.rgba *= color2.rgba;
+      gl_FragColor = color;
+    }
+    `,
+  mask: `
+    precision highp float;
+    uniform sampler2D uTexture;
+    uniform sampler2D uImage;
+    uniform vec4 uColor;
+    varying vec2 vTexCoord;
+    varying vec2 vTexCoord2;
+    void main() {
+      vec4 color = texture2D(uTexture, vTexCoord);
+      vec4 color2 = texture2D(uImage, vTexCoord2);
+      color.a = color2.a;
+      gl_FragColor = color;
+    }
+    `,
+} as const;
