+++ src/filters/gamma.shaders.ts
@@ -1,0 +1,15 @@
+export const fragmentSource = `
+  precision highp float;
+  uniform sampler2D uTexture;
+  uniform vec3 uGamma;
+  varying vec2 vTexCoord;
+  void main() {
+    vec4 color = texture2D(uTexture, vTexCoord);
+    vec3 correction = (1.0 / uGamma);
+    color.r = pow(color.r, correction.r);
+    color.g = pow(color.g, correction.g);
+    color.b = pow(color.b, correction.b);
+    gl_FragColor = color;
+    gl_FragColor.rgb *= color.a;
+  }
+`;
